+++ common/src/container/rle/png.rs
@@ -13,7 +13,7 @@
 //!   - <https://gist.github.com/juliusgeo/d4965b16a3c4478bb4eca2fe210559eb>
 
 use crate::{
-    container::{BitVec, Run},
+    container::{BitVec, Run, rle::intersperse_runs},
     serde::{DynamicSerializer, Serializer},
 };
 
@@ -78,7 +78,7 @@ impl<'a> PngEncoder<'a> {
 
     pub fn write_image_data(&mut self, mut data: Vec<Run>) {
         let width = self.size.x as u64 * self.color.planes() as u64;
-        intersperse_runs(&mut data, 0, width);
+        intersperse_runs(&mut data, Run::new(1, 0), 0, width);
 
         let mut check = Adler32::new();
         data.iter().for_each(|run| check.update_run(run));
@@ -103,59 +103,6 @@ impl ColorType {
         match self {
             ColorType::Grayscale => 1,
             ColorType::Truecolor => 3,
-        }
-    }
-}
-
-pub fn intersperse_runs(runs: &mut Vec<Run>, value: u8, spacing: u64) {
-    let mut i = 0; // The current run being processed
-    let mut pos = 0; // The current position in bytes
-    let mut next = 0; // Next byte index to insert `value`
-
-    while i < runs.len() {
-        let run = &mut runs[i];
-
-        // The range of positions covered by the current run. Excluding end.
-        // [start, pos)
-        let (start, end) = (pos, pos + run.length);
-
-        // If next insertion point is not in the range, advance to the next run.
-        // But if it is, split the run into parts left and right of the
-        // insertion point with the inserted run between.
-        if (start..end).contains(&next) {
-            // Avoid splitting run into parts if possible. When the values are
-            // the same, the length can just be updated.
-            if run.value == value {
-                let n = 1 + (end - next - 1) / spacing;
-                pos += run.length;
-                next += spacing * n;
-                run.length += n;
-                i += 1;
-            } else {
-                let run = runs.remove(i);
-
-                let length_left = next - start;
-                let length_right = run.length - length_left;
-                next += spacing;
-                pos += length_left;
-
-                if length_left > 0 {
-                    let (length, value) = (length_left, run.value);
-                    runs.insert(i, Run { length, value });
-                    i += 1;
-                }
-
-                runs.insert(i, Run { length: 1, value });
-                i += 1;
-
-                if length_right > 0 {
-                    let (length, value) = (length_right, run.value);
-                    runs.insert(i, Run { length, value });
-                }
-            }
-        } else {
-            pos += run.length;
-            i += 1;
         }
     }
 }